Updated and fixed function calls that lacked arguments in test classes.

+++ src/options.h
@@ -37,8 +37,8 @@ class Options {
   std::string regionsfile;
   std::string outprefix;
   // Insert sizes
-  int32_t dist_mean;
-  int32_t dist_sdev;
+  double dist_mean;
+  double dist_sdev;
   // Stutter model - TODO later make per locus model
   double stutter_up;
   double stutter_down;
